#95 - lib: catch SIGINT
