+++ src/lib/core/frame.rs
@@ -10,9 +10,8 @@
 use crate::{
     core::{
         apply::Core as _,
-        exception::{self, Condition, Exception},
+        exception::{self, Condition, Core as _, Exception},
         gc::Core as _,
-        lib::Core as _,
         mu::{Core as _, Mu},
         types::{Tag, Type},
     },
@@ -138,7 +137,7 @@ impl Frame {
 
     // apply
     pub fn apply(mut self, mu: &Mu, func: Tag) -> exception::Result<Tag> {
-        Mu::on_signal()?;
+        Exception::on_signal()?;
 
         match func.type_of() {
             Type::Symbol => {
